+++ Sandbox/GRCookbook/Chapter3/GL01_APIWrapping/src/GL.h
@@ -9,5 +9,10 @@ struct GL4API
 #	include "GLAPI.h"
 };
 
+namespace GL
+{
+    #include "GLAPI_.h"
+}
+
 void GetAPI4(GL4API* api, PFNGETGLPROC GetGLProc);
 void InjectAPITracer4(GL4API* api);
