+++ Sandbox/GRCookbook/Chapter3/GL01_APIWrapping/src/GLAPI_.h
@@ -1,0 +1,215 @@
+extern 	PFNGLACTIVETEXTUREPROC										glActiveTexture;
+extern 	PFNGLATTACHSHADERPROC										glAttachShader;
+extern 	PFNGLBEGINQUERYPROC											glBeginQuery;
+extern 	PFNGLBINDATTRIBLOCATIONPROC								glBindAttribLocation;
+extern 	PFNGLBINDBUFFERPROC											glBindBuffer;
+extern 	PFNGLBINDBUFFERBASEPROC										glBindBufferBase;
+extern 	PFNGLBINDBUFFERRANGEPROC									glBindBufferRange;
+extern 	PFNGLBINDFRAGDATALOCATIONPROC								glBindFragDataLocation;
+extern 	PFNGLBINDFRAMEBUFFERPROC									glBindFramebuffer;
+extern 	PFNGLBINDTEXTUREUNITPROC									glBindTextureUnit;
+extern 	PFNGLBINDTEXTURESPROC										glBindTextures;
+extern 	PFNGLBINDVERTEXARRAYPROC									glBindVertexArray;
+extern 	PFNGLBLENDFUNCPROC											glBlendFunc;
+extern 	PFNGLBLITNAMEDFRAMEBUFFERPROC								glBlitNamedFramebuffer;
+extern 	PFNGLBUFFERDATAPROC											glBufferData;
+extern 	PFNGLBUFFERSUBDATAPROC										glBufferSubData;
+extern 	PFNGLCHECKFRAMEBUFFERSTATUSPROC							glCheckFramebufferStatus;
+extern 	PFNGLCHECKNAMEDFRAMEBUFFERSTATUSPROC					glCheckNamedFramebufferStatus;
+extern 	PFNGLCLEARPROC													glClear;
+extern 	PFNGLCLEARCOLORPROC											glClearColor;
+extern 	PFNGLCLEARDEPTHPROC											glClearDepth;
+extern 	PFNGLCLEARNAMEDBUFFERDATAPROC								glClearNamedBufferData;
+extern 	PFNGLCLEARNAMEDBUFFERSUBDATAPROC							glClearNamedBufferSubData;
+extern 	PFNGLCLEARNAMEDFRAMEBUFFERFIPROC							glClearNamedFramebufferfi;
+extern 	PFNGLCLEARNAMEDFRAMEBUFFERFVPROC							glClearNamedFramebufferfv;
+extern 	PFNGLCLEARNAMEDFRAMEBUFFERIVPROC							glClearNamedFramebufferiv;
+extern 	PFNGLCLEARNAMEDFRAMEBUFFERUIVPROC						glClearNamedFramebufferuiv;
+extern 	PFNGLCLEARSTENCILPROC										glClearStencil;
+extern 	PFNGLCOLORMASKPROC											glColorMask;
+extern 	PFNGLCOMPILESHADERPROC										glCompileShader;
+extern 	PFNGLCOMPRESSEDTEXIMAGE2DPROC								glCompressedTexImage2D;
+extern 	PFNGLCOMPRESSEDTEXIMAGE3DPROC								glCompressedTexImage3D;
+extern 	PFNGLCOMPRESSEDTEXTURESUBIMAGE1DPROC					glCompressedTextureSubImage1D;
+extern 	PFNGLCOMPRESSEDTEXTURESUBIMAGE2DPROC					glCompressedTextureSubImage2D;
+extern 	PFNGLCOMPRESSEDTEXTURESUBIMAGE3DPROC					glCompressedTextureSubImage3D;
+extern 	PFNGLCOPYNAMEDBUFFERSUBDATAPROC							glCopyNamedBufferSubData;
+extern 	PFNGLCOPYTEXTURESUBIMAGE1DPROC							glCopyTextureSubImage1D;
+extern 	PFNGLCOPYTEXTURESUBIMAGE2DPROC							glCopyTextureSubImage2D;
+extern 	PFNGLCOPYTEXTURESUBIMAGE3DPROC							glCopyTextureSubImage3D;
+extern 	PFNGLCREATEBUFFERSPROC										glCreateBuffers;
+extern 	PFNGLCREATEFRAMEBUFFERSPROC								glCreateFramebuffers;
+extern 	PFNGLCREATEPROGRAMPROC										glCreateProgram;
+extern 	PFNGLCREATEPROGRAMPIPELINESPROC							glCreateProgramPipelines;
+extern 	PFNGLCREATEQUERIESPROC										glCreateQueries;
+extern 	PFNGLCREATERENDERBUFFERSPROC								glCreateRenderbuffers;
+extern 	PFNGLCREATESAMPLERSPROC										glCreateSamplers;
+extern 	PFNGLCREATESHADERPROC										glCreateShader;
+extern 	PFNGLCREATETEXTURESPROC										glCreateTextures;
+extern 	PFNGLCREATETRANSFORMFEEDBACKSPROC						glCreateTransformFeedbacks;
+extern 	PFNGLCREATEVERTEXARRAYSPROC								glCreateVertexArrays;
+extern 	PFNGLCULLFACEPROC												glCullFace;
+extern 	PFNGLDELETEBUFFERSPROC										glDeleteBuffers;
+extern 	PFNGLDELETEFRAMEBUFFERSPROC								glDeleteFramebuffers;
+extern 	PFNGLDELETEPROGRAMPROC										glDeleteProgram;
+extern 	PFNGLDELETEQUERIESPROC										glDeleteQueries;
+extern 	PFNGLDELETESHADERPROC										glDeleteShader;
+extern 	PFNGLDELETETEXTURESPROC										glDeleteTextures;
+extern 	PFNGLDELETEVERTEXARRAYSPROC								glDeleteVertexArrays;
+extern 	PFNGLDEPTHFUNCPROC											glDepthFunc;
+extern 	PFNGLDEPTHMASKPROC											glDepthMask;
+extern 	PFNGLDISABLEPROC												glDisable;
+extern 	PFNGLDISABLEVERTEXARRAYATTRIBPROC						glDisableVertexArrayAttrib;
+extern 	PFNGLDISABLEVERTEXATTRIBARRAYPROC						glDisableVertexAttribArray;
+extern 	PFNGLDISABLEIPROC												glDisablei;
+extern 	PFNGLDRAWARRAYSPROC											glDrawArrays;
+extern 	PFNGLDRAWARRAYSINSTANCEDPROC								glDrawArraysInstanced;
+extern 	PFNGLDRAWBUFFERSPROC											glDrawBuffers;
+extern 	PFNGLDRAWELEMENTSPROC										glDrawElements;
+extern 	PFNGLDRAWELEMENTSINSTANCEDPROC							glDrawElementsInstanced;
+extern 	PFNGLENABLEPROC												glEnable;
+extern 	PFNGLENABLEVERTEXARRAYATTRIBPROC							glEnableVertexArrayAttrib;
+extern 	PFNGLENABLEVERTEXATTRIBARRAYPROC							glEnableVertexAttribArray;
+extern 	PFNGLENABLEIPROC												glEnablei;
+extern 	PFNGLENDQUERYPROC												glEndQuery;
+extern 	PFNGLFINISHPROC												glFinish;
+extern 	PFNGLFLUSHPROC													glFlush;
+extern 	PFNGLFLUSHMAPPEDNAMEDBUFFERRANGEPROC					glFlushMappedNamedBufferRange;
+extern 	PFNGLFRAMEBUFFERTEXTURE2DPROC								glFramebufferTexture2D;
+extern 	PFNGLFRAMEBUFFERTEXTURE3DPROC								glFramebufferTexture3D;
+extern 	PFNGLGENBUFFERSPROC											glGenBuffers;
+extern 	PFNGLGENFRAMEBUFFERSPROC									glGenFramebuffers;
+extern 	PFNGLGENQUERIESPROC											glGenQueries;
+extern 	PFNGLGENTEXTURESPROC											glGenTextures;
+extern 	PFNGLGENVERTEXARRAYSPROC									glGenVertexArrays;
+extern 	PFNGLGENERATEMIPMAPPROC										glGenerateMipmap;
+extern 	PFNGLGENERATETEXTUREMIPMAPPROC							glGenerateTextureMipmap;
+extern 	PFNGLGETACTIVEATTRIBPROC									glGetActiveAttrib;
+extern 	PFNGLGETACTIVESUBROUTINENAMEPROC							glGetActiveSubroutineName;
+extern 	PFNGLGETACTIVESUBROUTINEUNIFORMIVPROC					glGetActiveSubroutineUniformiv;
+extern 	PFNGLGETACTIVEUNIFORMPROC									glGetActiveUniform;
+extern 	PFNGLGETACTIVEUNIFORMBLOCKIVPROC							glGetActiveUniformBlockiv;
+extern 	PFNGLGETATTRIBLOCATIONPROC									glGetAttribLocation;
+extern 	PFNGLGETCOMPRESSEDTEXIMAGEPROC							glGetCompressedTexImage;
+extern 	PFNGLGETCOMPRESSEDTEXTUREIMAGEPROC						glGetCompressedTextureImage;
+extern 	PFNGLGETERRORPROC												glGetError;
+extern 	PFNGLGETINTEGERVPROC											glGetIntegerv;
+extern 	PFNGLGETNAMEDBUFFERPARAMETERI64VPROC					glGetNamedBufferParameteri64v;
+extern 	PFNGLGETNAMEDBUFFERPARAMETERIVPROC						glGetNamedBufferParameteriv;
+extern 	PFNGLGETNAMEDBUFFERPOINTERVPROC							glGetNamedBufferPointerv;
+extern 	PFNGLGETNAMEDBUFFERSUBDATAPROC							glGetNamedBufferSubData;
+extern 	PFNGLGETNAMEDFRAMEBUFFERATTACHMENTPARAMETERIVPROC	glGetNamedFramebufferAttachmentParameteriv;
+extern 	PFNGLGETNAMEDFRAMEBUFFERPARAMETERIVPROC				glGetNamedFramebufferParameteriv;
+extern 	PFNGLGETNAMEDRENDERBUFFERPARAMETERIVPROC				glGetNamedRenderbufferParameteriv;
+extern 	PFNGLGETPROGRAMBINARYPROC									glGetProgramBinary;
+extern 	PFNGLGETPROGRAMINFOLOGPROC									glGetProgramInfoLog;
+extern 	PFNGLGETPROGRAMSTAGEIVPROC									glGetProgramStageiv;
+extern 	PFNGLGETPROGRAMIVPROC										glGetProgramiv;
+extern 	PFNGLGETQUERYOBJECTIVPROC									glGetQueryObjectiv;
+extern 	PFNGLGETQUERYOBJECTUI64VPROC								glGetQueryObjectui64v;
+extern 	PFNGLGETSHADERINFOLOGPROC									glGetShaderInfoLog;
+extern 	PFNGLGETSHADERIVPROC											glGetShaderiv;
+extern 	PFNGLGETSTRINGPROC											glGetString;
+extern 	PFNGLGETSTRINGIPROC											glGetStringi;
+extern 	PFNGLGETSUBROUTINEINDEXPROC								glGetSubroutineIndex;
+extern 	PFNGLGETSUBROUTINEUNIFORMLOCATIONPROC					glGetSubroutineUniformLocation;
+extern 	PFNGLGETTEXIMAGEPROC											glGetTexImage;
+extern 	PFNGLGETTEXLEVELPARAMETERIVPROC							glGetTexLevelParameteriv;
+extern 	PFNGLGETTEXTUREIMAGEPROC									glGetTextureImage;
+extern 	PFNGLGETTEXTURELEVELPARAMETERFVPROC						glGetTextureLevelParameterfv;
+extern 	PFNGLGETTEXTURELEVELPARAMETERIVPROC						glGetTextureLevelParameteriv;
+extern 	PFNGLGETTEXTUREPARAMETERIIVPROC							glGetTextureParameterIiv;
+extern 	PFNGLGETTEXTUREPARAMETERIUIVPROC							glGetTextureParameterIuiv;
+extern 	PFNGLGETTEXTUREPARAMETERFVPROC							glGetTextureParameterfv;
+extern 	PFNGLGETTEXTUREPARAMETERIVPROC							glGetTextureParameteriv;
+extern 	PFNGLGETTRANSFORMFEEDBACKI64_VPROC						glGetTransformFeedbacki64_v;
+extern 	PFNGLGETTRANSFORMFEEDBACKI_VPROC							glGetTransformFeedbacki_v;
+extern 	PFNGLGETTRANSFORMFEEDBACKIVPROC							glGetTransformFeedbackiv;
+extern 	PFNGLGETUNIFORMLOCATIONPROC								glGetUniformLocation;
+extern 	PFNGLGETVERTEXARRAYINDEXED64IVPROC						glGetVertexArrayIndexed64iv;
+extern 	PFNGLGETVERTEXARRAYINDEXEDIVPROC							glGetVertexArrayIndexediv;
+extern 	PFNGLGETVERTEXARRAYIVPROC									glGetVertexArrayiv;
+extern 	PFNGLINVALIDATENAMEDFRAMEBUFFERDATAPROC				glInvalidateNamedFramebufferData;
+extern 	PFNGLINVALIDATENAMEDFRAMEBUFFERSUBDATAPROC			glInvalidateNamedFramebufferSubData;
+extern 	PFNGLISPROGRAMPROC											glIsProgram;
+extern 	PFNGLISSHADERPROC												glIsShader;
+extern 	PFNGLLINKPROGRAMPROC											glLinkProgram;
+extern 	PFNGLMAPNAMEDBUFFERPROC										glMapNamedBuffer;
+extern 	PFNGLMAPNAMEDBUFFERRANGEPROC								glMapNamedBufferRange;
+extern 	PFNGLNAMEDBUFFERDATAPROC									glNamedBufferData;
+extern 	PFNGLNAMEDBUFFERSTORAGEPROC								glNamedBufferStorage;
+extern 	PFNGLNAMEDBUFFERSUBDATAPROC								glNamedBufferSubData;
+extern 	PFNGLNAMEDFRAMEBUFFERDRAWBUFFERPROC						glNamedFramebufferDrawBuffer;
+extern 	PFNGLNAMEDFRAMEBUFFERDRAWBUFFERSPROC					glNamedFramebufferDrawBuffers;
+extern 	PFNGLNAMEDFRAMEBUFFERPARAMETERIPROC						glNamedFramebufferParameteri;
+extern 	PFNGLNAMEDFRAMEBUFFERREADBUFFERPROC						glNamedFramebufferReadBuffer;
+extern 	PFNGLNAMEDFRAMEBUFFERRENDERBUFFERPROC					glNamedFramebufferRenderbuffer;
+extern 	PFNGLNAMEDFRAMEBUFFERTEXTUREPROC							glNamedFramebufferTexture;
+extern 	PFNGLNAMEDFRAMEBUFFERTEXTURELAYERPROC					glNamedFramebufferTextureLayer;
+extern 	PFNGLNAMEDRENDERBUFFERSTORAGEPROC						glNamedRenderbufferStorage;
+extern 	PFNGLNAMEDRENDERBUFFERSTORAGEMULTISAMPLEPROC			glNamedRenderbufferStorageMultisample;
+extern 	PFNGLPATCHPARAMETERIPROC									glPatchParameteri;
+extern 	PFNGLPIXELSTOREIPROC											glPixelStorei;
+extern 	PFNGLPOLYGONMODEPROC											glPolygonMode;
+extern 	PFNGLPROGRAMBINARYPROC										glProgramBinary;
+extern 	PFNGLPROGRAMPARAMETERIPROC									glProgramParameteri;
+extern 	PFNGLPROGRAMUNIFORM1FPROC									glProgramUniform1f;
+extern 	PFNGLPROGRAMUNIFORM1IPROC									glProgramUniform1i;
+extern 	PFNGLPROGRAMUNIFORM2FVPROC									glProgramUniform2fv;
+extern 	PFNGLPROGRAMUNIFORM2IVPROC									glProgramUniform2iv;
+extern 	PFNGLPROGRAMUNIFORM3FVPROC									glProgramUniform3fv;
+extern 	PFNGLPROGRAMUNIFORM3IVPROC									glProgramUniform3iv;
+extern 	PFNGLPROGRAMUNIFORM4FVPROC									glProgramUniform4fv;
+extern 	PFNGLPROGRAMUNIFORM4IVPROC									glProgramUniform4iv;
+extern 	PFNGLREADBUFFERPROC											glReadBuffer;
+extern 	PFNGLREADPIXELSPROC											glReadPixels;
+extern 	PFNGLSCISSORPROC												glScissor;
+extern 	PFNGLSHADERSOURCEPROC										glShaderSource;
+extern 	PFNGLTEXIMAGE2DPROC											glTexImage2D;
+extern 	PFNGLTEXIMAGE3DPROC											glTexImage3D;
+extern 	PFNGLTEXPARAMETERFPROC										glTexParameterf;
+extern 	PFNGLTEXPARAMETERFVPROC										glTexParameterfv;
+extern 	PFNGLTEXPARAMETERIPROC										glTexParameteri;
+extern 	PFNGLTEXPARAMETERIVPROC										glTexParameteriv;
+extern 	PFNGLTEXSUBIMAGE2DPROC										glTexSubImage2D;
+extern 	PFNGLTEXTUREBUFFERPROC										glTextureBuffer;
+extern 	PFNGLTEXTUREBUFFERRANGEPROC								glTextureBufferRange;
+extern 	PFNGLTEXTUREPARAMETERIIVPROC								glTextureParameterIiv;
+extern 	PFNGLTEXTUREPARAMETERIUIVPROC								glTextureParameterIuiv;
+extern 	PFNGLTEXTUREPARAMETERFPROC									glTextureParameterf;
+extern 	PFNGLTEXTUREPARAMETERFVPROC								glTextureParameterfv;
+extern 	PFNGLTEXTUREPARAMETERIPROC									glTextureParameteri;
+extern 	PFNGLTEXTUREPARAMETERIVPROC								glTextureParameteriv;
+extern 	PFNGLTEXTURESTORAGE1DPROC									glTextureStorage1D;
+extern 	PFNGLTEXTURESTORAGE2DPROC									glTextureStorage2D;
+extern 	PFNGLTEXTURESTORAGE2DMULTISAMPLEPROC					glTextureStorage2DMultisample;
+extern 	PFNGLTEXTURESTORAGE3DPROC									glTextureStorage3D;
+extern 	PFNGLTEXTURESTORAGE3DMULTISAMPLEPROC					glTextureStorage3DMultisample;
+extern 	PFNGLTEXTURESUBIMAGE1DPROC									glTextureSubImage1D;
+extern 	PFNGLTEXTURESUBIMAGE2DPROC									glTextureSubImage2D;
+extern 	PFNGLTEXTURESUBIMAGE3DPROC									glTextureSubImage3D;
+extern 	PFNGLTRANSFORMFEEDBACKBUFFERBASEPROC					glTransformFeedbackBufferBase;
+extern 	PFNGLTRANSFORMFEEDBACKBUFFERRANGEPROC					glTransformFeedbackBufferRange;
+extern 	PFNGLUNIFORM1FPROC											glUniform1f;
+extern 	PFNGLUNIFORM1FVPROC											glUniform1fv;
+extern 	PFNGLUNIFORM1IPROC											glUniform1i;
+extern 	PFNGLUNIFORM1IVPROC											glUniform1iv;
+extern 	PFNGLUNIFORM3FVPROC											glUniform3fv;
+extern 	PFNGLUNIFORM4FVPROC											glUniform4fv;
+extern 	PFNGLUNIFORMBLOCKBINDINGPROC								glUniformBlockBinding;
+extern 	PFNGLUNIFORMMATRIX3FVPROC									glUniformMatrix3fv;
+extern 	PFNGLUNIFORMMATRIX4FVPROC									glUniformMatrix4fv;
+extern 	PFNGLUNIFORMSUBROUTINESUIVPROC							glUniformSubroutinesuiv;
+extern 	PFNGLUNMAPNAMEDBUFFERPROC									glUnmapNamedBuffer;
+extern 	PFNGLUSEPROGRAMPROC											glUseProgram;
+extern 	PFNGLVALIDATEPROGRAMPROC									glValidateProgram;
+extern 	PFNGLVERTEXARRAYATTRIBBINDINGPROC						glVertexArrayAttribBinding;
+extern 	PFNGLVERTEXARRAYATTRIBFORMATPROC							glVertexArrayAttribFormat;
+extern 	PFNGLVERTEXARRAYATTRIBIFORMATPROC						glVertexArrayAttribIFormat;
+extern 	PFNGLVERTEXARRAYATTRIBLFORMATPROC						glVertexArrayAttribLFormat;
+extern 	PFNGLVERTEXARRAYBINDINGDIVISORPROC						glVertexArrayBindingDivisor;
+extern 	PFNGLVERTEXARRAYELEMENTBUFFERPROC						glVertexArrayElementBuffer;
+extern 	PFNGLVERTEXARRAYVERTEXBUFFERPROC							glVertexArrayVertexBuffer;
+extern 	PFNGLVERTEXARRAYVERTEXBUFFERSPROC						glVertexArrayVertexBuffers;
+extern 	PFNGLVERTEXATTRIBPOINTERPROC								glVertexAttribPointer;
+extern 	PFNGLVIEWPORTPROC												glViewport;
